WIIS-63: mise en place rôles et accès pour admin + stock + collecte

+++ src/Controller/CollecteController.php
@@ -17,6 +17,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 
 use App\Repository\CollecteRepository;
 use App\Repository\ArticleRepository;
@@ -197,28 +198,32 @@ class CollecteController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $em = $this->getDoctrine()->getEntityManager();
-            $date = new \DateTime('now');
-            $status = $this->statutRepository->findOneByNom(Collecte::STATUS_DEMANDE);
-            $numero = "C-" . $date->format('YmdHis');
-            $collecte = new Collecte;
-            $collecte
-                ->setDemandeur($this->utilisateurRepository->find($data['demandeur']))
-                ->setNumero($numero)
-                ->setDate($date)
-                ->setStatut($status)
-                ->setPointCollecte($this->emplacementRepository->find($data['Pcollecte']))
-                ->setObjet($data['Objet'])
-                ->setCommentaire($data['commentaire']);
-            $em->persist($collecte);
-            $em->flush();
-            $data = [
-                "redirect" => $this->generateUrl('collecte_show', ['id' => $collecte->getId()])
-            ];
-            return new JsonResponse($data);
+        if ($this->isGranted('ROLE_SUPER_CUSTOMER')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $em = $this->getDoctrine()->getEntityManager();
+                $date = new \DateTime('now');
+                $status = $this->statutRepository->findOneByNom(Collecte::STATUS_DEMANDE);
+                $numero = "C-" . $date->format('YmdHis');
+                $collecte = new Collecte;
+                $collecte
+                    ->setDemandeur($this->utilisateurRepository->find($data['demandeur']))
+                    ->setNumero($numero)
+                    ->setDate($date)
+                    ->setStatut($status)
+                    ->setPointCollecte($this->emplacementRepository->find($data['Pcollecte']))
+                    ->setObjet($data['Objet'])
+                    ->setCommentaire($data['commentaire']);
+                $em->persist($collecte);
+                $em->flush();
+                $data = [
+                    "redirect" => $this->generateUrl('collecte_show', ['id' => $collecte->getId()])
+                ];
+                return new JsonResponse($data);
+            }
+            throw new XmlHttpException("404 not found");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new XmlHttpException("404 not found");
     }
 
 
@@ -331,29 +336,33 @@ class CollecteController extends AbstractController
      */
     public function edit(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $collecte = $this->collecteRepository->find($data['collecte']);
-            $pointCollecte = $this->emplacementRepository->find($data['Pcollecte']);
+        if ($this->isGranted('ROLE_SUPER_CUSTOMER')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $collecte = $this->collecteRepository->find($data['collecte']);
+                $pointCollecte = $this->emplacementRepository->find($data['Pcollecte']);
 
-            $collecte
-                ->setNumero($data["NumeroCollecte"])
-                ->setDate(new \DateTime($data["date-collecte"]))
-                ->setCommentaire($data["commentaire"])
-                ->setObjet($data["objet"])
-                ->setPointCollecte($pointCollecte);
+                $collecte
+                    ->setNumero($data["NumeroCollecte"])
+                    ->setDate(new \DateTime($data["date-collecte"]))
+                    ->setCommentaire($data["commentaire"])
+                    ->setObjet($data["objet"])
+                    ->setPointCollecte($pointCollecte);
 
-            $em = $this->getDoctrine()->getManager();
-            $em->flush();
-            $json = [
-                'entete' => $this->renderView('collecte/enteteCollecte.html.twig', [
-                    'collecte' => $collecte,
-                    'modifiable' => ($collecte->getStatut()->getNom() !== Collecte::STATUS_EN_COURS ?true : false)
-                ])
-            ];
+                $em = $this->getDoctrine()->getManager();
+                $em->flush();
+                $json = [
+                    'entete' => $this->renderView('collecte/enteteCollecte.html.twig', [
+                        'collecte' => $collecte,
+                        'modifiable' => ($collecte->getStatut()->getNom() !== Collecte::STATUS_EN_COURS ? true : false)
+                    ])
+                ];
 
-            return new JsonResponse($json);
+                return new JsonResponse($json);
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
@@ -361,16 +370,20 @@ class CollecteController extends AbstractController
      */
     public function delete(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $collecte = $this->collecteRepository->find($data['collecte']);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($collecte);
-            $entityManager->flush();
-            $data = [
-                "redirect" => $this->generateUrl('collecte_index')
-            ];
-            return new JsonResponse($data);
+        if ($this->isGranted('ROLE_ADMIN_GT')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $collecte = $this->collecteRepository->find($data['collecte']);
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($collecte);
+                $entityManager->flush();
+                $data = [
+                    "redirect" => $this->generateUrl('collecte_index')
+                ];
+                return new JsonResponse($data);
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 }
